+++ src/main/java/caculator/models/Calculator.java
@@ -1,23 +1,21 @@
 package caculator.models;
 
 import caculator.enums.Operator;
+import caculator.utils.Util;
 
 import java.util.List;
 import java.util.Stack;
 
 public class Calculator {
 
-    private String[] formula;
-    private List<String> postfixFormula = new Stack<>();
-    private Stack<String> operands = new Stack<>();
+    private List<String> postfixFormula;
     private Stack<Integer> numbers = new Stack<>();
 
-    public Calculator(String[] formula) {
-        this.formula = formula;
+    public Calculator(List<String> postfixFormula) {
+        this.postfixFormula = postfixFormula;
     }
 
     public Integer calculate() {
-        convertPostfixFormula();
         for(String postfixFormulaStr: postfixFormula) {
             addNumberInNumbers(postfixFormulaStr);
             addCalculatedNumberInNumbers(postfixFormulaStr);
@@ -28,12 +26,12 @@ public class Calculator {
     }
 
     private void addNumberInNumbers(String postfixFormulaStr) {
-        if (isNumber(postfixFormulaStr))
+        if (Util.isNumber(postfixFormulaStr))
             numbers.push(Integer.parseInt(postfixFormulaStr));
     }
 
     private void addCalculatedNumberInNumbers(String postfixFormulaStr) {
-        if (isOperator(postfixFormulaStr)) {
+        if (Util.isOperator(postfixFormulaStr)) {
             checkNumbersHasTwoNumber();
             Integer secondOperand = numbers.pop();
             Integer firstOperand = numbers.pop();
@@ -47,44 +45,4 @@ public class Calculator {
             throw new RuntimeException("올바르지 않은 계산식입니다.");
     }
 
-    private void convertPostfixFormula() {
-        for(String formulaStr: formula) {
-            checkIncorrectFormula(formulaStr);
-            addNumberInPostfixFormula(formulaStr);
-            addOperatorInPostfixFormula(formulaStr);
-        }
-        while(!operands.isEmpty()) {
-            postfixFormula.add(operands.pop());
-        }
-    }
-
-    private void addNumberInPostfixFormula(String formulaStr) {
-        if(isNumber(formulaStr))
-            postfixFormula.add(formulaStr);
-    }
-
-    private void addOperatorInPostfixFormula(String formulaStr) {
-        if(isOperator(formulaStr)) {
-            if(!operands.isEmpty()) {
-                postfixFormula.add(operands.pop());
-            }
-            operands.push(formulaStr);
-        }
-    }
-
-    private void checkIncorrectFormula(String formulaStr) {
-        if(!isNumber(formulaStr) && !isOperator(formulaStr))
-            throw new RuntimeException("올바르지 않은 계산식입니다.");
-    }
-
-    private boolean isOperator(String operator) {
-        return operator.equals(Operator.PLUS.getOperator())
-                || operator.equals(Operator.MINUS.getOperator())
-                || operator.equals(Operator.MULTIPLY.getOperator())
-                || operator.equals(Operator.DIVIDE.getOperator());
-    }
-
-    private boolean isNumber(String number) {
-        return number.chars().allMatch(Character::isDigit);
-    }
 }
